% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composeCompanies.R
\name{composeCompanies}
\alias{composeCompanies}
\title{Compose the database for Companies}
\usage{
composeCompanies(dataPath = "/../data/companies/")
}
\arguments{
\item{dataPath:}{path of raw data}
}
\value{
The Companies dataset
}
\description{
Compose the database for Companies
}
