% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composePeople.R
\name{composePeople}
\alias{composePeople}
\title{Compose the database for Companies}
\usage{
composePeople(dataPath = "/../data/_people/")
}
\arguments{
\item{dataPath:}{path of raw data}
}
\value{
The People dataset
}
\description{
Compose the database for Companies
}
