% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFounders.R
\name{readFounders}
\alias{readFounders}
\title{Compose the database for Companies}
\usage{
readFounders(dataPath = "/../data/people/")
}
\arguments{
\item{dataPath:}{path of raw data}
}
\value{
The People dataset
}
\description{
Compose the database for Companies
}
