\docType{package}
\name{Rpad-package}
\alias{Rpad-package}
\title{Rpad}
\description{
  Workbook-style, web-based interface to R
}
\details{
This manual describes the R functions that support the web interface.
Other documentation is in the \dQuote{basehtml} directory.
The example below will launch a browser where you can read that documentation and view the demos.
You can continue to use the R console after the local HTTP server starts.
}
\examples{
\dontrun{
# launches the local version
Rpad()
}
}
\keyword{package}

